requirements.txt
matplotlib
pandas
python-binance
python-dotenv